% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{handle_affectations}
\alias{handle_affectations}
\title{Function used by server to handle affectations}
\usage{
handle_affectations(input, output, df)
}
\arguments{
\item{input}{Input data from the frontend}

\item{output}{Output data the frontend will receive}

\item{df}{The reactive data frame of students's wishes and affectations}
}
\description{
Function used by server to handle affectations
}
