% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{select_file_type}
\alias{select_file_type}
\title{Reads the file and selects the needed columns}
\usage{
select_file_type(file_path)
}
\arguments{
\item{file_path}{The path to the file to parse}
}
\value{
The file as a data frame
}
\description{
Reads the file and selects the needed columns
}
