% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Frontend UI layout}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
ui
}
\description{
Frontend UI layout
}
\keyword{datasets}
